fix build errors; refactor
